Issue-179 : size_t type byte arrays

+++ api-tests/dev_apis/crypto/test_c058/test_data.h
@@ -34,7 +34,7 @@ typedef struct {
     size_t                  plaintext_length;
     size_t                  input_length;
     size_t                  output_size;
-    size_t                  expected_output[45];
+    uint8_t                 expected_output[45];
     size_t                  expected_length;
     psa_status_t            expected_status;
 } test_data;
